requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
PyYAML==6.0
requests==2.28.1
schedule==1.1.0
soupsieve==2.3.2.post1
urllib3==1.26.10
